Fixed output directory inconsistencies across all files
The "." link to a directory's self does not specify a consistent starting location for an output directory. While specifying something like "./Output" would work if you have already navigated to the preferred directory through the console, we want to ensure that the output directory for the program is consistent regardless of what the user has done prior to running the program.

Adding the powershell specific variable "$PSScriptRoot" solves this problem by consistently navigating to the directory of the executing parent script and forcing output files to drop to this location.

In this commit I have replaced all occurrences of "./" with "$PSScriptRoot" in an attempt to solve output directory inconsistency. All functions have been tested after this change and all functions have been confirmed working, as well as their corresponding output directory.