update recycle bin
since verifying recycle bin contents seems to break the program, I'll leave that as a stretch goal. The original code is commented out but can be implemented with some tweaks and a little more powershell know-how. this function returns true regardless of whether or not the files are copied correctly. This will be noted in the tooltip